Move around modules so things are drop in compatible and all the specialized (an unspecialized) versions are available
